+++ python/sglang/test/kits/gsm8k_accuracy_kit.py
@@ -1,13 +1,18 @@
 from types import SimpleNamespace
+from typing import Optional
+
+import requests
 
 from sglang.test.few_shot_gsm8k import run_eval as run_eval_gsm8k
-from sglang.test.test_utils import CustomTestCase
 
 
 class GSM8KMixin:
     accuracy: float
+    spec_accept_length_thres: Optional[float] = None
 
-    def test_gsm8k(self: CustomTestCase):
+    def test_gsm8k(self):
+        requests.get(self.base_url + "/flush_cache")
+
         args = SimpleNamespace(
             num_shots=5,
             data_path=None,
@@ -20,3 +25,11 @@ class GSM8KMixin:
         metrics = run_eval_gsm8k(args)
         print(f"{metrics=}")
         self.assertGreaterEqual(metrics["accuracy"], self.accuracy)
+
+        if self.spec_accept_length_thres is not None:
+            server_info = requests.get(self.base_url + "/server_info")
+            avg_spec_accept_length = server_info.json()["internal_states"][0][
+                "avg_spec_accept_length"
+            ]
+            print(f"{avg_spec_accept_length=}")
+            self.assertGreater(avg_spec_accept_length, self.spec_accept_length_thres)
